ask2live/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.1
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==3.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-mysql==3.11.1
django-redis==4.12.1
django-sslserver==0.22
djangorestframework==3.12.2
drf-yasg==1.20.0
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.1.0
hiredis==1.1.0
hpack==3.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.0
mysqlclient==2.0.3
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
priority==1.3.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
websocket==0.2.1
websocket-client==0.58.0
websockets==8.1
zope.event==4.5.0
zope.interface==5.3.0
